Restructure ts-components package
