+++ packages/ts-components/src/index.ts
@@ -1,2 +1,2 @@
-export { default as ViewCountWrapper } from './ViewCountWrapper';
-export { default as InArticlePuff } from './InArticlePuff';
+export { default as ViewCountWrapper } from './view-count-wrapper/ViewCountWrapper';
+export { default as InArticlePuff } from './in-article-puff';
